+++ src/includes/replies/functions.php
@@ -1166,6 +1166,7 @@ function bbp_get_reply_ancestors( $reply_id = 0 ) {
 			$ancestors = array( $reply_to );
 
 			// Get parent reply
+			// phpcs:ignore Generic.CodeAnalysis.AssignmentInCondition.FoundInWhileCondition -- Fetch each ancestor until none remains.
 			while ( $ancestor = bbp_get_reply( $id ) ) {
 
 				// Does parent have a parent?
@@ -2467,10 +2468,15 @@ function bbp_get_reply_position_raw( $reply_id = 0, $topic_id = 0 ) {
 
 				// Reverse replies array and search for current reply position
 				$topic_replies  = array_reverse( $topic_replies );
-				$reply_position = array_search( (string) $reply_id, $topic_replies );
+				$reply_position = array_search( $reply_id, $topic_replies, true );
 
-				// Bump the position to compensate for the lead topic post
-				++$reply_position;
+				// Bump the position to compensate for the lead topic post,
+				// or reset to 0 if the reply was not found in the children list.
+				if ( false !== $reply_position ) {
+					++$reply_position;
+				} else {
+					$reply_position = 0;
+				}
 			}
 		}
 	}